+++ mayan/apps/dynamic_search/backends/elasticsearch.py
@@ -5,8 +5,6 @@ import elasticsearch
 from elasticsearch import Elasticsearch, helpers
 from elasticsearch_dsl import Q, Search, analyzer, char_filter
 
-from mayan.apps.common.utils import parse_range
-
 from ..classes import SearchBackend, SearchModel
 from ..exceptions import DynamicSearchException
 from ..settings import setting_results_limit
@@ -204,17 +202,14 @@ class ElasticSearchBackend(SearchBackend):
             id=instance.pk, document=document
         )
 
-    def index_search_model(self, search_model, range_string=None):
+    def index_instances(self, search_model, id_list):
         client = self.get_client()
         index_name = self.get_index_name(search_model=search_model)
 
         def generate_actions():
             queryset = search_model.get_queryset()
 
-            if range_string:
-                queryset = queryset.filter(
-                    pk__in=list(parse_range(range_string=range_string))
-                )
+            queryset = queryset.filter(pk__in=id_list)
 
             for instance in queryset:
                 kwargs = search_model.populate(