+++ mayan/apps/dynamic_search/queues.py
@@ -21,9 +21,9 @@ queue_search.add_task_type(
     name='task_index_instance',
 )
 queue_search.add_task_type(
-    dotted_path='mayan.apps.dynamic_search.tasks.task_index_search_model',
+    dotted_path='mayan.apps.dynamic_search.tasks.task_index_instances',
     label=_('Index all instances of a search model to the search engine.'),
-    name='task_index_search_model',
+    name='task_index_instances',
 )
 queue_search_slow.add_task_type(
     dotted_path='mayan.apps.dynamic_search.tasks.task_reindex_backend',