+++ mayan/apps/dynamic_search/tasks.py
@@ -1,7 +1,6 @@
 import logging
 
 from django.apps import apps
-from django.db.models.aggregates import Max, Min
 
 from mayan.apps.lock_manager.exceptions import LockError
 from mayan.celery import app
@@ -9,9 +8,9 @@ from mayan.celery import app
 from .classes import SearchBackend, SearchModel
 from .exceptions import DynamicSearchException, DynamicSearchRetry
 from .literals import (
-    TASK_DEINDEX_INSTANCE_MAX_RETRIES, TASK_INDEX_INSTANCE_MAX_RETRIES
+    TASK_DEINDEX_INSTANCE_MAX_RETRIES, TASK_INDEX_INSTANCE_MAX_RETRIES,
+    TASK_INDEX_INSTANCES_MAX_RETRIES
 )
-from .settings import setting_indexing_chunk_size
 
 logger = logging.getLogger(name=__name__)
 
@@ -32,25 +31,6 @@ def task_deindex_instance(self, app_label, model_name, object_id):
         raise self.retry(exc=exception)
 
     logger.info('Finished')
-
-
-@app.task(
-    bind=True, ignore_result=True, max_retries=None, retry_backoff=True
-)
-def task_index_search_model(self, search_model_full_name, range_string=None):
-    search_model = SearchModel.get(name=search_model_full_name)
-
-    kwargs = {
-        'range_string': range_string, 'search_model': search_model
-    }
-
-    try:
-        SearchBackend.get_instance().index_search_model(**kwargs)
-    except Exception as exception:
-        raise DynamicSearchException(
-            'Unexpected error calling `index_search_model` with '
-            'keyword arguments {}.'.format(kwargs)
-        ) from exception
 
 
 @app.task(
@@ -80,8 +60,49 @@ def task_index_instance(
         )
     except (DynamicSearchRetry, LockError) as exception:
         raise self.retry(exc=exception)
+    except Exception as exception:
+        kwargs = {
+            'app_label': app_label,
+            'model_name': model_name,
+            'object_id': object_id,
+            'exclude_app_label': exclude_app_label,
+            'exclude_model_name': exclude_model_name,
+            'exclude_kwargs': exclude_kwargs
+        }
+        error_message = (
+            'Unexpected error calling `task_index_instance` with keyword '
+            'arguments {}.'
+        ).format(kwargs)
 
-    logger.info('Finished')
+        logger.error(error_message)
+        raise DynamicSearchException(error_message) from exception
+    else:
+        logger.info('Finished')
+
+
+@app.task(
+    bind=True, ignore_result=True,
+    max_retries=TASK_INDEX_INSTANCES_MAX_RETRIES, retry_backoff=True
+)
+def task_index_instances(self, search_model_full_name, id_list):
+    search_model = SearchModel.get(name=search_model_full_name)
+
+    kwargs = {
+        'id_list': id_list, 'search_model': search_model
+    }
+
+    try:
+        SearchBackend.get_instance().index_instances(**kwargs)
+    except (DynamicSearchRetry, LockError) as exception:
+        raise self.retry(exc=exception)
+    except Exception as exception:
+        error_message = (
+            'Unexpected error calling `task_index_instances` with '
+            'keyword arguments {}.'
+        ).format(kwargs)
+
+        logger.error(error_message)
+        raise DynamicSearchException(error_message) from exception
 
 
 @app.task(ignore_result=True)
@@ -90,16 +111,10 @@ def task_reindex_backend():
     backend.reset()
 
     for search_model in SearchModel.all():
-        queryset = search_model.model._meta.managers_map[search_model.manager_name].all()
-        if queryset:
-            queryset = queryset.aggregate(min_id=Min('id'), max_id=Max('id'))
-            step = setting_indexing_chunk_size.value - 1
-
-            for id_start in range(queryset['min_id'], queryset['max_id'] + 1, step):
-                range_string = '{}-{}'.format(id_start, id_start + step - 1)
-                task_index_search_model.apply_async(
-                    kwargs={
-                        'range_string': range_string,
-                        'search_model_full_name': search_model.get_full_name(),
-                    }
-                )
+        for id_list in search_model.get_id_groups():
+            task_index_instances.apply_async(
+                kwargs={
+                    'id_list': id_list,
+                    'search_model_full_name': search_model.get_full_name(),
+                }
+            )